Revamped the Trigger system
This allows multiple triggers to be added with conditions.
The triggers are specified in an enum so users can have ease of access when producing their triggers.

The conditions however still need knowledge of the basic advancement system. Later on, this could also be made easier to use.

The icon method also now accepts a material value.